Implement Helm-/TailorDeploymentStrategies
This pulls out the logic of the deployments in respective strategies to get rid of the intertwined logic between Helm and Tailor.
