+++ src/org/ods/component/IDeploymentStrategy.groovy
@@ -1,0 +1,8 @@
+package org.ods.component
+
+import org.ods.util.PodData
+
+@SuppressWarnings('MethodCount')
+interface IDeploymentStrategy {
+    Map<String, List<PodData>> deploy()
+}
